feat(trip): trip api to return trip information
