+++ fasta-backend/api/transporters-api.js
@@ -1,46 +1,34 @@
-/* eslint-disable max-len */
-/* eslint-disable no-console */
-/* eslint-disable class-methods-use-this */
-/* eslint-disable no-undef */
-/* eslint-disable no-unused-vars */
-const https = require("https");
+const { Client, Status } = require("@googlemaps/google-maps-services-js");
 
-// Option parameter for google api request
-const options = {
-  hostname: "maps.googleapis.com",
-  port: 443,
-  path: `/maps/api/place/textsearch/json?query=transport&location=${this.latitude},${this.longitude}&radius=10000&key=AIzaSyC8G-WQhDTvlVflIFPzOj30rDQGk0Fekx8`,
-  method: "GET"
-};
+// instantiate the google http client
+const client = new Client({});
 
 // Allow function to receive coordinate argument from client
-function GetTransporter(latitude, longitude) {
+function GetTripInfo(latitude, longitude) {
   this.latitude = latitude;
   this.longitude = longitude;
 }
 
-// COmpute polygon
-GetTransporter.prototype.computePolygon = () => {
-// empty function
-// const google = new google.maps
-};
 
-// Get nearby transport 1km away
-GetTransporter.prototype.getPlaces = async () => {
-  const req = await https.request(options, (res) => {
-    console.log(`statusCode: ${res.statusCode}`);
-    res.on("data", (d) => {
-      process.stdout.write(d);
+// api to list all the transporters within 1km of users location
+GetTripInfo.prototype.getPlaces = async () => {
+  client.textSearch({
+    params: {
+      locations: [{ lat: this.latitude, lng: this.longitude }],
+      query: "transport",
+      key: process.env.TEST_KEY,
+      radius: 10000
+    }
+  })
+    .then((r) => {
+      if (r.data.status === Status.OK) {
+        return r.data.results;
+      }
+      return r.data.error_message;
+    })
+    .catch((e) => {
+      console.log(e);
     });
-  });
-  req.on("error", (error) => {
-    console.error(error);
-  });
-
-  req.end();
 };
 
-//
-
-
-module.exports = GetTransporter;
+module.exports = GetTripInfo;
